services/requirements.txt
fastapi
uvicorn
SQLAlchemy
SQLite4
redis
boto3
requests
psutil
GPUtil
numpy
opencv-python-headless
python-multipart
mlflow
hydra-core
hydra-optuna-sweeper
optuna
wsqlite
wredis
loguru
logbook
